+++ adc/adc.c
@@ -179,7 +179,7 @@ void DMA_Tx_Init(DMA_Channel_TypeDef *DMA_CHx, u32 ppadr, u32 memadr, u16 bufsiz
     NVIC_Init(&NVIC_InitStructure); // ≈‰÷√NVIC
 }
 
-void show_battery(UWORD Xpoint, UWORD Ypoint, UWORD Color_Background, UWORD Color_Foreground, int *needshow)
+void show_battery(UWORD Xpoint, UWORD Ypoint, UWORD Color_Background, UWORD Color_Foreground, u8 *needshow)
 {
 
     static u8 percentage = 0;
