+++ screen/screen_api.h
@@ -42,7 +42,7 @@ extern LCD_0IN85_ATTRIBUTES LCD;
 extern u8 dmaXpoint,dmaYpoint;
 extern sFONT* dmaFont;
 extern volatile int dma_circular;
-extern volatile int dmaTransferComplete ;
+extern volatile u8 dmaTransferComplete ;
